requirements.txt
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
python-dateutil==2.8.2